refactor: split editor_widget.go into feature files
editor_widget.go had grown to ~2780 lines. Relocate cohesive method
clusters into same-package files: multicursor, line ops, word/case text
ops, bracket matching/colorization, and pure helpers. Main file drops to
~1570 lines (struct, constructor, Render, HandleEvent, core plumbing).

Pure mechanical move, no API or behavior change.
