+++ internal/ui/editor_widget.go
@@ -1,6 +1,11 @@
 package ui
 
 import (
+	"strconv"
+	"strings"
+	"time"
+	"unicode"
+
 	"github.com/eugenioenko/ttt/internal/core/buffer"
 	"github.com/eugenioenko/ttt/internal/core/cursor"
 	"github.com/eugenioenko/ttt/internal/core/diff"
@@ -11,11 +16,6 @@ import (
 	"github.com/eugenioenko/ttt/internal/core/undo"
 	"github.com/eugenioenko/ttt/internal/term"
 	"github.com/eugenioenko/ttt/internal/view"
-	"sort"
-	"strconv"
-	"strings"
-	"time"
-	"unicode"
 
 	"github.com/gdamore/tcell/v2"
 )
@@ -1228,32 +1228,6 @@ func (e *EditorPaneWidget) mouseToPos(r Rect, mx, my int) (line, col int) {
 	return
 }
 
-// wordAt returns the maximal run of word characters (letters, digits, or '_')
-// surrounding col in lineText, or "" if col is not on a word character.
-func wordAt(lineText string, col int) string {
-	runes := []rune(lineText)
-	if len(runes) == 0 || col < 0 {
-		return ""
-	}
-	if col >= len(runes) {
-		col = len(runes) - 1
-	}
-	isWord := func(r rune) bool {
-		return unicode.IsLetter(r) || unicode.IsDigit(r) || r == '_'
-	}
-	if !isWord(runes[col]) {
-		return ""
-	}
-	start, end := col, col
-	for start > 0 && isWord(runes[start-1]) {
-		start--
-	}
-	for end < len(runes)-1 && isWord(runes[end+1]) {
-		end++
-	}
-	return string(runes[start : end+1])
-}
-
 func (e *EditorPaneWidget) selectWord(line, col int) {
 	if e.Selection == nil || line < 0 || line >= len(e.Buf.Lines) {
 		return
@@ -1340,65 +1314,6 @@ func (e *EditorPaneWidget) dedentForCloser() {
 	}
 }
 
-func leadingIndentWidth(line string, tabSize int) int {
-	runes := []rune(line)
-	if len(runes) > 0 && runes[0] == '\t' {
-		return 1
-	}
-	remove := 0
-	for remove < tabSize && remove < len(runes) && runes[remove] == ' ' {
-		remove++
-	}
-	return remove
-}
-
-func leadingWhitespace(s string) string {
-	for i, r := range s {
-		if r != ' ' && r != '\t' {
-			return s[:i]
-		}
-	}
-	return s
-}
-
-func bufColToVisualCol(line string, bufCol, tabW int) int {
-	visCol := 0
-	ri := 0
-	for _, ch := range line {
-		if ri >= bufCol {
-			break
-		}
-		if ch == '\t' {
-			visCol = ((visCol / tabW) + 1) * tabW
-		} else {
-			visCol++
-		}
-		ri++
-	}
-	return visCol
-}
-
-func visualColToBufCol(line string, targetVisCol, tabW int) int {
-	visCol := 0
-	ri := 0
-	for _, ch := range line {
-		if visCol >= targetVisCol {
-			return ri
-		}
-		if ch == '\t' {
-			nextStop := ((visCol / tabW) + 1) * tabW
-			if targetVisCol < nextStop {
-				return ri
-			}
-			visCol = nextStop
-		} else {
-			visCol++
-		}
-		ri++
-	}
-	return len([]rune(line))
-}
-
 type screenCell struct {
 	ch     rune
 	style  term.Style
@@ -1450,182 +1365,6 @@ func (e *EditorPaneWidget) renderLineToScreen(line []rune, spans []highlight.Spa
 		}
 	}
 	return cells
-}
-
-var bracketPairs = map[rune]rune{
-	'(': ')', ')': '(',
-	'[': ']', ']': '[',
-	'{': '}', '}': '{',
-}
-
-var closingBrackets = map[rune]bool{')': true, ']': true, '}': true}
-
-var indentOpeners = map[rune]bool{'{': true, '(': true, '[': true, ':': true}
-
-func (e *EditorPaneWidget) findMatchingBracket() (int, int, bool) {
-	if e.Cursor.Line < 0 || e.Cursor.Line >= len(e.Buf.Lines) {
-		return 0, 0, false
-	}
-	runes := []rune(e.Buf.Lines[e.Cursor.Line])
-	if e.Cursor.Col < 0 || e.Cursor.Col >= len(runes) {
-		return 0, 0, false
-	}
-	ch := runes[e.Cursor.Col]
-	match, ok := bracketPairs[ch]
-	if !ok {
-		return 0, 0, false
-	}
-	dir := 1
-	if closingBrackets[ch] {
-		dir = -1
-	}
-	depth := 1
-	line, col := e.Cursor.Line, e.Cursor.Col
-	for {
-		col += dir
-		lr := []rune(e.Buf.Lines[line])
-		if col < 0 {
-			line--
-			if line < 0 {
-				return 0, 0, false
-			}
-			lr = []rune(e.Buf.Lines[line])
-			col = len(lr) - 1
-			if col < 0 {
-				col = 0
-				continue
-			}
-		} else if col >= len(lr) {
-			line++
-			if line >= len(e.Buf.Lines) {
-				return 0, 0, false
-			}
-			col = 0
-			lr = []rune(e.Buf.Lines[line])
-			if len(lr) == 0 {
-				continue
-			}
-		}
-		c := lr[col]
-		if c == ch {
-			depth++
-		} else if c == match {
-			depth--
-			if depth == 0 {
-				return line, col, true
-			}
-		}
-	}
-}
-
-func (e *EditorPaneWidget) GoToMatchingBracket() {
-	line, col, ok := e.findMatchingBracket()
-	if ok {
-		e.ExpandFoldContaining(line)
-		e.Cursor.Line = line
-		e.Cursor.Col = col
-		if e.Selection != nil && e.Selection.Active {
-			e.Selection.Clear()
-		}
-		e.scrollViewport()
-	}
-}
-
-var openBrackets = map[rune]bool{'(': true, '[': true, '{': true}
-
-type bracketColorEntry struct {
-	col   int
-	style term.Style
-}
-
-type bracketColorMap [][]bracketColorEntry
-
-func (e *EditorPaneWidget) computeBracketColors() bracketColorMap {
-	totalLines := len(e.Buf.Lines)
-	if len(e.BracketColorStyles) == 0 || totalLines == 0 {
-		return nil
-	}
-
-	result := make(bracketColorMap, totalLines)
-	bracketStyles := e.BracketColorStyles
-	numStyles := len(bracketStyles)
-	depth := 0
-
-	inLineComment := false
-	inBlockComment := false
-	inString := false
-	var stringChar rune
-
-	for lineIdx := 0; lineIdx < totalLines; lineIdx++ {
-		runes := []rune(e.Buf.Lines[lineIdx])
-		inLineComment = false
-		n := len(runes)
-
-		for i := 0; i < n; i++ {
-			ch := runes[i]
-			next := rune(0)
-			if i+1 < n {
-				next = runes[i+1]
-			}
-
-			if inBlockComment {
-				if ch == '*' && next == '/' {
-					inBlockComment = false
-					i++
-				}
-				continue
-			}
-			if inLineComment {
-				continue
-			}
-			if inString {
-				if ch == '\\' {
-					i++
-					continue
-				}
-				if ch == stringChar {
-					inString = false
-				}
-				continue
-			}
-
-			if ch == '/' && next == '/' {
-				inLineComment = true
-				continue
-			}
-			if ch == '/' && next == '*' {
-				inBlockComment = true
-				i++
-				continue
-			}
-			if ch == '"' || ch == '\'' || ch == '`' {
-				inString = true
-				stringChar = ch
-				if ch == '`' {
-					// backtick strings can span lines; handled by not resetting inString
-				}
-				continue
-			}
-
-			if openBrackets[ch] {
-				style := bracketStyles[depth%numStyles]
-				depth++
-				result[lineIdx] = append(result[lineIdx], bracketColorEntry{col: i, style: style})
-			} else if closingBrackets[ch] {
-				depth--
-				if depth < 0 {
-					depth = 0
-				}
-				style := bracketStyles[depth%numStyles]
-				result[lineIdx] = append(result[lineIdx], bracketColorEntry{col: i, style: style})
-			}
-		}
-
-		if inString && stringChar != '`' {
-			inString = false
-		}
-	}
-	return result
 }
 
 func (e *EditorPaneWidget) clampCursor() {
@@ -1743,10 +1482,6 @@ func (e *EditorPaneWidget) scrollViewportWrap() {
 	}
 }
 
-func isEditorIdentRune(r rune) bool {
-	return r == '_' || (r >= 'a' && r <= 'z') || (r >= 'A' && r <= 'Z') || (r >= '0' && r <= '9')
-}
-
 func (e *EditorPaneWidget) scrollUp(n int) {
 	if e.WordWrap {
 		tabW := e.resolveTabSize()
@@ -1822,435 +1557,6 @@ func (e *EditorPaneWidget) scrollDown(n int) {
 	}
 }
 
-func (e *EditorPaneWidget) MoveLineUp() {
-	hasSel := e.Selection != nil && e.Selection.Active
-	if hasSel {
-		start, end := e.Selection.Range(e.Cursor.Line, e.Cursor.Col)
-		if start.Line <= 0 {
-			return
-		}
-		for line := start.Line; line <= end.Line; line++ {
-			e.exec(&undo.SwapLineCommand{Line1: line, Line2: line - 1})
-		}
-		e.Cursor.Line--
-		e.Selection.Anchor.Line--
-	} else {
-		if e.Cursor.Line <= 0 {
-			return
-		}
-		e.exec(&undo.SwapLineCommand{Line1: e.Cursor.Line, Line2: e.Cursor.Line - 1})
-		e.Cursor.Line--
-	}
-	e.clampCursor()
-	e.scrollViewport()
-}
-
-func (e *EditorPaneWidget) MoveLineDown() {
-	hasSel := e.Selection != nil && e.Selection.Active
-	if hasSel {
-		start, end := e.Selection.Range(e.Cursor.Line, e.Cursor.Col)
-		if end.Line >= len(e.Buf.Lines)-1 {
-			return
-		}
-		for line := end.Line; line >= start.Line; line-- {
-			e.exec(&undo.SwapLineCommand{Line1: line, Line2: line + 1})
-		}
-		e.Cursor.Line++
-		e.Selection.Anchor.Line++
-	} else {
-		if e.Cursor.Line >= len(e.Buf.Lines)-1 {
-			return
-		}
-		e.exec(&undo.SwapLineCommand{Line1: e.Cursor.Line, Line2: e.Cursor.Line + 1})
-		e.Cursor.Line++
-	}
-	e.clampCursor()
-	e.scrollViewport()
-}
-
-func (e *EditorPaneWidget) DuplicateLine() {
-	e.Cursor.Line = e.Buf.ClampLine(e.Cursor.Line)
-	text := e.Buf.Lines[e.Cursor.Line]
-	e.exec(&undo.InsertLineCommand{Idx: e.Cursor.Line + 1, Text: text})
-	e.Cursor.Line++
-	e.clampCursor()
-	e.scrollViewport()
-}
-
-func (e *EditorPaneWidget) DeleteLine() {
-	e.Cursor.Line = e.Buf.ClampLine(e.Cursor.Line)
-	if len(e.Buf.Lines) <= 1 {
-		e.exec(&undo.DeleteSelectionCommand{
-			StartLine: 0, StartCol: 0,
-			EndLine: 0, EndCol: len([]rune(e.Buf.Lines[0])),
-		})
-		e.Cursor.Col = 0
-	} else {
-		e.exec(&undo.DeleteLineCommand{Idx: e.Cursor.Line})
-		e.Cursor.Line = e.Buf.ClampLine(e.Cursor.Line)
-	}
-	lineLen := len([]rune(e.Buf.Lines[e.Cursor.Line]))
-	if e.Cursor.Col > lineLen {
-		e.Cursor.Col = lineLen
-	}
-	e.clampCursor()
-	e.scrollViewport()
-}
-
-func (e *EditorPaneWidget) JoinLines() {
-	if e.Undo != nil {
-		e.Undo.BreakGroup()
-	}
-	hasSel := e.Selection != nil && e.Selection.Active
-	if hasSel {
-		start, end := e.Selection.Range(e.Cursor.Line, e.Cursor.Col)
-		endLine := end.Line
-		if end.Col == 0 && endLine > start.Line {
-			endLine--
-		}
-		// Join all lines in the selection range from the first line
-		for endLine > start.Line {
-			cmd := &undo.JoinNextLineCommand{Line: start.Line}
-			e.exec(cmd)
-			endLine--
-		}
-		e.Selection.Active = false
-		e.Cursor.Line = start.Line
-		lineLen := len([]rune(e.Buf.Lines[start.Line]))
-		if e.Cursor.Col > lineLen {
-			e.Cursor.Col = lineLen
-		}
-	} else {
-		if e.Cursor.Line >= len(e.Buf.Lines)-1 {
-			return
-		}
-		cmd := &undo.JoinNextLineCommand{Line: e.Cursor.Line}
-		e.exec(cmd)
-		e.Cursor.Col = cmd.JoinCol
-	}
-	e.clampCursor()
-	e.scrollViewport()
-}
-
-func (e *EditorPaneWidget) InsertLineBelow() {
-	e.exec(&undo.InsertLineCommand{Idx: e.Cursor.Line + 1, Text: ""})
-	e.Cursor.Line++
-	e.Cursor.Col = 0
-	e.clampCursor()
-	e.scrollViewport()
-}
-
-func (e *EditorPaneWidget) InsertLineAbove() {
-	e.exec(&undo.InsertLineCommand{Idx: e.Cursor.Line, Text: ""})
-	e.Cursor.Col = 0
-	e.clampCursor()
-	e.scrollViewport()
-}
-
-func (e *EditorPaneWidget) commentPrefix() string {
-	prefix := "//"
-	if e.Highlighter != nil {
-		lang := strings.ToLower(e.Highlighter.Language())
-		switch lang {
-		case "python", "ruby", "bash", "shell", "yaml", "toml":
-			prefix = "#"
-		case "lua", "sql":
-			prefix = "--"
-		case "html", "xml":
-			prefix = "<!--"
-		}
-	}
-	return prefix
-}
-
-func (e *EditorPaneWidget) ToggleLineComment() {
-	prefix := e.commentPrefix()
-
-	startLine, endLine := e.Cursor.Line, e.Cursor.Line
-	if e.Selection != nil && e.Selection.Active {
-		start, end := e.Selection.Range(e.Cursor.Line, e.Cursor.Col)
-		startLine = start.Line
-		endLine = end.Line
-		if end.Col == 0 && endLine > startLine {
-			endLine--
-		}
-	}
-	startLine = e.Buf.ClampLine(startLine)
-	endLine = e.Buf.ClampLine(endLine)
-
-	allCommented := true
-	for l := startLine; l <= endLine; l++ {
-		trimmed := strings.TrimLeft(e.Buf.Lines[l], " \t")
-		if trimmed == "" {
-			continue
-		}
-		if !strings.HasPrefix(trimmed, prefix) {
-			allCommented = false
-			break
-		}
-	}
-
-	var cmds []undo.EditCommand
-	cursorDelta := 0
-
-	for l := startLine; l <= endLine; l++ {
-		runes := []rune(e.Buf.Lines[l])
-		trimmed := strings.TrimLeft(string(runes), " \t")
-		if trimmed == "" {
-			continue
-		}
-		indent := len(runes) - len([]rune(trimmed))
-
-		if allCommented {
-			removeLen := len([]rune(prefix))
-			if indent+removeLen < len(runes) && runes[indent+removeLen] == ' ' {
-				removeLen++
-			}
-			cmd := &undo.DeleteSelectionCommand{
-				StartLine: l, StartCol: indent,
-				EndLine: l, EndCol: indent + removeLen,
-			}
-			cmd.Apply(e.Buf)
-			cmds = append(cmds, cmd)
-			if l == e.Cursor.Line {
-				cursorDelta = -removeLen
-			}
-		} else {
-			cmd := &undo.InsertStringCommand{Line: l, Col: indent, Text: prefix + " "}
-			cmd.Apply(e.Buf)
-			cmds = append(cmds, cmd)
-			if l == e.Cursor.Line {
-				cursorDelta = len([]rune(prefix)) + 1
-			}
-		}
-	}
-
-	if len(cmds) > 0 && e.Undo != nil {
-		e.Undo.Push(&undo.BatchCommand{Commands: cmds})
-		e.bufferDirty = true
-	}
-
-	e.Cursor.Col += cursorDelta
-	if e.Cursor.Col < 0 {
-		e.Cursor.Col = 0
-	}
-	e.clampCursor()
-	e.scrollViewport()
-}
-
-// lineRange returns the start and end line indices for the current selection,
-// or the full buffer range if no selection is active.
-func (e *EditorPaneWidget) lineRange() (int, int) {
-	if e.Selection != nil && e.Selection.Active {
-		start, end := e.Selection.Range(e.Cursor.Line, e.Cursor.Col)
-		endLine := end.Line
-		if end.Col == 0 && endLine > start.Line {
-			endLine--
-		}
-		return e.Buf.ClampLine(start.Line), e.Buf.ClampLine(endLine)
-	}
-	end := len(e.Buf.Lines) - 1
-	if end > 0 && e.Buf.Lines[end] == "" {
-		end--
-	}
-	return 0, end
-}
-
-// copyLines returns a copy of the buffer lines in the given range (inclusive).
-func (e *EditorPaneWidget) copyLines(start, end int) []string {
-	lines := make([]string, end-start+1)
-	copy(lines, e.Buf.Lines[start:end+1])
-	return lines
-}
-
-func (e *EditorPaneWidget) SortLinesAsc() {
-	start, end := e.lineRange()
-	old := e.copyLines(start, end)
-	sorted := make([]string, len(old))
-	copy(sorted, old)
-	sort.Strings(sorted)
-	e.exec(&undo.ReplaceLinesCommand{Start: start, OldLines: old, NewLines: sorted})
-	e.clampCursor()
-	e.scrollViewport()
-}
-
-func (e *EditorPaneWidget) SortLinesDesc() {
-	start, end := e.lineRange()
-	old := e.copyLines(start, end)
-	sorted := make([]string, len(old))
-	copy(sorted, old)
-	sort.Sort(sort.Reverse(sort.StringSlice(sorted)))
-	e.exec(&undo.ReplaceLinesCommand{Start: start, OldLines: old, NewLines: sorted})
-	e.clampCursor()
-	e.scrollViewport()
-}
-
-func (e *EditorPaneWidget) ReverseLines() {
-	start, end := e.lineRange()
-	old := e.copyLines(start, end)
-	reversed := make([]string, len(old))
-	for i, line := range old {
-		reversed[len(old)-1-i] = line
-	}
-	e.exec(&undo.ReplaceLinesCommand{Start: start, OldLines: old, NewLines: reversed})
-	e.clampCursor()
-	e.scrollViewport()
-}
-
-func (e *EditorPaneWidget) UniqueLines() {
-	start, end := e.lineRange()
-	old := e.copyLines(start, end)
-	seen := make(map[string]bool)
-	var unique []string
-	for _, line := range old {
-		if !seen[line] {
-			seen[line] = true
-			unique = append(unique, line)
-		}
-	}
-	e.exec(&undo.ReplaceLinesCommand{Start: start, OldLines: old, NewLines: unique})
-	e.clampCursor()
-	e.scrollViewport()
-}
-
-func wordBoundaryLeft(runes []rune, col int) int {
-	pos := col - 1
-	if pos >= len(runes) {
-		pos = len(runes) - 1
-	}
-	if pos < 0 {
-		return 0
-	}
-	if unicode.IsSpace(runes[pos]) {
-		for pos > 0 && unicode.IsSpace(runes[pos-1]) {
-			pos--
-		}
-	} else if isEditorIdentRune(runes[pos]) {
-		for pos > 0 && isEditorIdentRune(runes[pos-1]) {
-			pos--
-		}
-	} else {
-		for pos > 0 && !isEditorIdentRune(runes[pos-1]) && !unicode.IsSpace(runes[pos-1]) {
-			pos--
-		}
-	}
-	return pos
-}
-
-func wordBoundaryRight(runes []rune, col int) int {
-	pos := col
-	if pos >= len(runes) {
-		return len(runes)
-	}
-	if unicode.IsSpace(runes[pos]) {
-		for pos < len(runes) && unicode.IsSpace(runes[pos]) {
-			pos++
-		}
-	} else if isEditorIdentRune(runes[pos]) {
-		for pos < len(runes) && isEditorIdentRune(runes[pos]) {
-			pos++
-		}
-	} else {
-		for pos < len(runes) && !isEditorIdentRune(runes[pos]) && !unicode.IsSpace(runes[pos]) {
-			pos++
-		}
-	}
-	return pos
-}
-
-func (e *EditorPaneWidget) MoveWordLeft(shift bool) {
-	if e.isMultiActive() && !shift {
-		e.multiMoveAll(func(cs *multicursor.CursorState) {
-			if cs.Col == 0 {
-				if cs.Line > 0 {
-					cs.Line--
-					cs.Col = len([]rune(e.Buf.Lines[cs.Line]))
-				}
-			} else {
-				runes := []rune(e.Buf.Lines[cs.Line])
-				cs.Col = wordBoundaryLeft(runes, cs.Col)
-			}
-		})
-		e.scrollViewport()
-		return
-	}
-	e.startOrExtendSelection(shift)
-	e.Cursor.Line = e.Buf.ClampLine(e.Cursor.Line)
-	if e.Cursor.Col == 0 {
-		if e.Cursor.Line > 0 {
-			e.Cursor.Line--
-			e.Cursor.Col = len([]rune(e.Buf.Lines[e.Cursor.Line]))
-		}
-	} else {
-		runes := []rune(e.Buf.Lines[e.Cursor.Line])
-		e.Cursor.Col = wordBoundaryLeft(runes, e.Cursor.Col)
-	}
-	e.clampCursor()
-	e.scrollViewport()
-}
-
-func (e *EditorPaneWidget) MoveWordRight(shift bool) {
-	if e.isMultiActive() && !shift {
-		e.multiMoveAll(func(cs *multicursor.CursorState) {
-			runes := []rune(e.Buf.Lines[cs.Line])
-			if cs.Col >= len(runes) {
-				if cs.Line < len(e.Buf.Lines)-1 {
-					cs.Line++
-					cs.Col = 0
-				}
-			} else {
-				cs.Col = wordBoundaryRight(runes, cs.Col)
-			}
-		})
-		e.scrollViewport()
-		return
-	}
-	e.startOrExtendSelection(shift)
-	e.Cursor.Line = e.Buf.ClampLine(e.Cursor.Line)
-	runes := []rune(e.Buf.Lines[e.Cursor.Line])
-	if e.Cursor.Col >= len(runes) {
-		if e.Cursor.Line < len(e.Buf.Lines)-1 {
-			e.Cursor.Line++
-			e.Cursor.Col = 0
-		}
-	} else {
-		e.Cursor.Col = wordBoundaryRight(runes, e.Cursor.Col)
-	}
-	e.clampCursor()
-	e.scrollViewport()
-}
-
-func (e *EditorPaneWidget) DeleteWordLeft() {
-	if e.Cursor.Col == 0 {
-		return
-	}
-	e.Cursor.Line = e.Buf.ClampLine(e.Cursor.Line)
-	runes := []rune(e.Buf.Lines[e.Cursor.Line])
-	start := wordBoundaryLeft(runes, e.Cursor.Col)
-	e.exec(&undo.DeleteSelectionCommand{
-		StartLine: e.Cursor.Line, StartCol: start,
-		EndLine: e.Cursor.Line, EndCol: e.Cursor.Col,
-	})
-	e.Cursor.Col = start
-	e.clampCursor()
-	e.scrollViewport()
-}
-
-func (e *EditorPaneWidget) DeleteWordRight() {
-	e.Cursor.Line = e.Buf.ClampLine(e.Cursor.Line)
-	runes := []rune(e.Buf.Lines[e.Cursor.Line])
-	if e.Cursor.Col >= len(runes) {
-		return
-	}
-	end := wordBoundaryRight(runes, e.Cursor.Col)
-	e.exec(&undo.DeleteSelectionCommand{
-		StartLine: e.Cursor.Line, StartCol: e.Cursor.Col,
-		EndLine: e.Cursor.Line, EndCol: end,
-	})
-	e.clampCursor()
-	e.scrollViewport()
-}
-
 func (e *EditorPaneWidget) SmartHome() {
 	runes := []rune(e.Buf.Lines[e.Cursor.Line])
 	firstNonSpace := 0
@@ -2264,519 +1570,4 @@ func (e *EditorPaneWidget) SmartHome() {
 	}
 	e.clampCursor()
 	e.scrollViewport()
-}
-
-func (e *EditorPaneWidget) ensureMulti() {
-	if e.Multi == nil {
-		e.Multi = multicursor.New(e.Cursor.Line, e.Cursor.Col)
-		if e.Selection != nil && e.Selection.Active {
-			e.Multi.Cursors[0].Sel = *e.Selection
-		}
-	}
-}
-
-func (e *EditorPaneWidget) syncFromMulti() {
-	if e.Multi == nil || len(e.Multi.Cursors) == 0 {
-		return
-	}
-	p := e.Multi.PrimaryCursor()
-	e.Cursor.Line = p.Line
-	e.Cursor.Col = p.Col
-	if e.Selection != nil {
-		*e.Selection = p.Sel
-	}
-}
-
-func (e *EditorPaneWidget) syncToMulti() {
-	if e.Multi == nil || len(e.Multi.Cursors) == 0 {
-		return
-	}
-	c := &e.Multi.Cursors[e.Multi.Primary]
-	c.Line = e.Cursor.Line
-	c.Col = e.Cursor.Col
-	if e.Selection != nil {
-		c.Sel = *e.Selection
-	}
-}
-
-func (e *EditorPaneWidget) isMultiActive() bool {
-	return e.Multi != nil && e.Multi.IsMulti()
-}
-
-func (e *EditorPaneWidget) collapseMulti() {
-	if e.Multi == nil {
-		return
-	}
-	e.Multi.CollapseToSingle()
-	e.syncFromMulti()
-	e.Multi = nil
-	e.multiSearchWord = ""
-}
-
-func (e *EditorPaneWidget) multiExecRune(r rune) {
-	e.syncToMulti()
-	var cmds []undo.EditCommand
-	for i := len(e.Multi.Cursors) - 1; i >= 0; i-- {
-		cs := &e.Multi.Cursors[i]
-		if cs.Sel.Active {
-			start, end := cs.Sel.Range(cs.Line, cs.Col)
-			delCmd := &undo.DeleteSelectionCommand{
-				StartLine: start.Line, StartCol: start.Col,
-				EndLine: end.Line, EndCol: end.Col,
-			}
-			delCmd.Apply(e.Buf)
-			cmds = append(cmds, delCmd)
-			cs.Line = start.Line
-			cs.Col = start.Col
-			cs.Sel.Clear()
-			e.adjustLaterCursors(i, start, end)
-		}
-		insertCol := cs.Col
-		cmd := &undo.InsertRuneCommand{Line: cs.Line, Col: insertCol, Rune: r}
-		cmd.Apply(e.Buf)
-		cmds = append(cmds, cmd)
-		cs.Col++
-		e.shiftLaterCursors(i, cs.Line, insertCol, 1)
-	}
-	if e.Undo != nil {
-		e.Undo.Push(&undo.BatchCommand{Commands: cmds})
-	}
-	e.bufferDirty = true
-	e.syncFromMulti()
-}
-
-func (e *EditorPaneWidget) multiExecBackspace() {
-	e.syncToMulti()
-	var cmds []undo.EditCommand
-	for i := len(e.Multi.Cursors) - 1; i >= 0; i-- {
-		cs := &e.Multi.Cursors[i]
-		if cs.Sel.Active {
-			start, end := cs.Sel.Range(cs.Line, cs.Col)
-			delCmd := &undo.DeleteSelectionCommand{
-				StartLine: start.Line, StartCol: start.Col,
-				EndLine: end.Line, EndCol: end.Col,
-			}
-			delCmd.Apply(e.Buf)
-			cmds = append(cmds, delCmd)
-			cs.Line = start.Line
-			cs.Col = start.Col
-			cs.Sel.Clear()
-			e.adjustLaterCursors(i, start, end)
-			continue
-		}
-		cs.Line = e.Buf.ClampLine(cs.Line)
-		if cs.Col > 0 {
-			cmd := &undo.DeleteRuneCommand{Line: cs.Line, Col: cs.Col - 1}
-			cmd.Apply(e.Buf)
-			cmds = append(cmds, cmd)
-			cs.Col--
-			e.shiftLaterCursors(i, cs.Line, cs.Col, -1)
-		} else if cs.Line > 0 {
-			prevLen := len([]rune(e.Buf.Lines[cs.Line-1]))
-			cmd := &undo.JoinLineCommand{Line: cs.Line}
-			cmd.Apply(e.Buf)
-			cmds = append(cmds, cmd)
-			e.shiftLaterLines(i, cs.Line, -1)
-			cs.Line--
-			cs.Col = prevLen
-		}
-	}
-	if len(cmds) > 0 {
-		if e.Undo != nil {
-			e.Undo.Push(&undo.BatchCommand{Commands: cmds})
-		}
-		e.bufferDirty = true
-	}
-	e.Multi.Deduplicate()
-	e.syncFromMulti()
-}
-
-func (e *EditorPaneWidget) multiExecDelete() {
-	e.syncToMulti()
-	var cmds []undo.EditCommand
-	for i := len(e.Multi.Cursors) - 1; i >= 0; i-- {
-		cs := &e.Multi.Cursors[i]
-		if cs.Sel.Active {
-			start, end := cs.Sel.Range(cs.Line, cs.Col)
-			delCmd := &undo.DeleteSelectionCommand{
-				StartLine: start.Line, StartCol: start.Col,
-				EndLine: end.Line, EndCol: end.Col,
-			}
-			delCmd.Apply(e.Buf)
-			cmds = append(cmds, delCmd)
-			cs.Line = start.Line
-			cs.Col = start.Col
-			cs.Sel.Clear()
-			e.adjustLaterCursors(i, start, end)
-			continue
-		}
-		cs.Line = e.Buf.ClampLine(cs.Line)
-		lineLen := len([]rune(e.Buf.Lines[cs.Line]))
-		if cs.Col < lineLen {
-			cmd := &undo.DeleteRuneCommand{Line: cs.Line, Col: cs.Col}
-			cmd.Apply(e.Buf)
-			cmds = append(cmds, cmd)
-			e.shiftLaterCursors(i, cs.Line, cs.Col, -1)
-		} else if cs.Line < len(e.Buf.Lines)-1 {
-			cmd := &undo.JoinLineCommand{Line: cs.Line + 1}
-			cmd.Apply(e.Buf)
-			cmds = append(cmds, cmd)
-			e.shiftLaterLines(i, cs.Line+1, -1)
-		}
-	}
-	if len(cmds) > 0 {
-		if e.Undo != nil {
-			e.Undo.Push(&undo.BatchCommand{Commands: cmds})
-		}
-		e.bufferDirty = true
-	}
-	e.Multi.Deduplicate()
-	e.syncFromMulti()
-}
-
-func (e *EditorPaneWidget) multiExecEnter() {
-	e.syncToMulti()
-	var cmds []undo.EditCommand
-	for i := len(e.Multi.Cursors) - 1; i >= 0; i-- {
-		cs := &e.Multi.Cursors[i]
-		if cs.Sel.Active {
-			start, end := cs.Sel.Range(cs.Line, cs.Col)
-			delCmd := &undo.DeleteSelectionCommand{
-				StartLine: start.Line, StartCol: start.Col,
-				EndLine: end.Line, EndCol: end.Col,
-			}
-			delCmd.Apply(e.Buf)
-			cmds = append(cmds, delCmd)
-			cs.Line = start.Line
-			cs.Col = start.Col
-			cs.Sel.Clear()
-			e.adjustLaterCursors(i, start, end)
-		}
-		cs.Line = e.Buf.ClampLine(cs.Line)
-		indent := leadingWhitespace(e.Buf.Lines[cs.Line])
-		cmd := &undo.SplitLineCommand{Line: cs.Line, Col: cs.Col}
-		cmd.Apply(e.Buf)
-		cmds = append(cmds, cmd)
-		e.shiftLaterLines(i, cs.Line, 1)
-		cs.Line++
-		cs.Col = 0
-		if len(indent) > 0 {
-			indCmd := &undo.InsertStringCommand{Line: cs.Line, Col: 0, Text: indent}
-			indCmd.Apply(e.Buf)
-			cmds = append(cmds, indCmd)
-			cs.Col = len([]rune(indent))
-		}
-	}
-	if e.Undo != nil {
-		e.Undo.Push(&undo.BatchCommand{Commands: cmds})
-	}
-	e.bufferDirty = true
-	e.syncFromMulti()
-}
-
-func (e *EditorPaneWidget) adjustLaterCursors(editedIdx int, start, end selection.Position) {
-	for j := editedIdx + 1; j < len(e.Multi.Cursors); j++ {
-		cs := &e.Multi.Cursors[j]
-		if start.Line == end.Line {
-			if cs.Line == start.Line && cs.Col >= end.Col {
-				cs.Col -= end.Col - start.Col
-			}
-		} else {
-			if cs.Line == end.Line {
-				cs.Col = start.Col + (cs.Col - end.Col)
-				cs.Line = start.Line
-			} else if cs.Line > end.Line {
-				cs.Line -= end.Line - start.Line
-			}
-		}
-	}
-}
-
-func (e *EditorPaneWidget) shiftLaterCursors(editedIdx, line, col, delta int) {
-	for j := editedIdx + 1; j < len(e.Multi.Cursors); j++ {
-		cs := &e.Multi.Cursors[j]
-		if cs.Line == line && cs.Col >= col {
-			cs.Col += delta
-			if cs.Col < 0 {
-				cs.Col = 0
-			}
-		}
-	}
-}
-
-func (e *EditorPaneWidget) shiftLaterLines(editedIdx, fromLine, delta int) {
-	for j := editedIdx + 1; j < len(e.Multi.Cursors); j++ {
-		if e.Multi.Cursors[j].Line >= fromLine {
-			e.Multi.Cursors[j].Line += delta
-		}
-	}
-}
-
-func (e *EditorPaneWidget) multiMoveAll(moveFn func(cs *multicursor.CursorState)) {
-	e.syncToMulti()
-	for i := range e.Multi.Cursors {
-		e.Multi.Cursors[i].Sel.Clear()
-		e.Multi.Cursors[i].Line = e.Buf.ClampLine(e.Multi.Cursors[i].Line)
-		moveFn(&e.Multi.Cursors[i])
-		e.Multi.Cursors[i].Line = e.Buf.ClampLine(e.Multi.Cursors[i].Line)
-	}
-	e.Multi.Deduplicate()
-	e.syncFromMulti()
-}
-
-func (e *EditorPaneWidget) SelectNextOccurrence() {
-	if e.Selection == nil {
-		return
-	}
-	word := ""
-	if e.Selection.Active {
-		word = e.Selection.Text(e.Buf.Lines, e.Cursor.Line, e.Cursor.Col)
-	}
-	if word == "" {
-		e.selectWord(e.Cursor.Line, e.Cursor.Col)
-		if e.Selection.Active {
-			e.multiSearchWord = e.Selection.Text(e.Buf.Lines, e.Cursor.Line, e.Cursor.Col)
-		}
-		e.ensureMulti()
-		e.syncToMulti()
-		return
-	}
-	if e.multiSearchWord == "" {
-		e.multiSearchWord = word
-	}
-	e.ensureMulti()
-	e.syncToMulti()
-	e.Multi.NormalizePrimary()
-	e.syncFromMulti()
-
-	searchWord := e.multiSearchWord
-	lastCursor := e.Multi.Cursors[len(e.Multi.Cursors)-1]
-	startLine := lastCursor.Line
-	startCol := lastCursor.Col
-
-	for line := startLine; line < len(e.Buf.Lines)+startLine; line++ {
-		l := line % len(e.Buf.Lines)
-		runes := []rune(e.Buf.Lines[l])
-		searchRunes := []rune(searchWord)
-		fromCol := 0
-		if l == startLine {
-			fromCol = startCol
-		}
-		for col := fromCol; col <= len(runes)-len(searchRunes); col++ {
-			if string(runes[col:col+len(searchRunes)]) == searchWord {
-				already := false
-				for _, c := range e.Multi.Cursors {
-					s, end := c.Sel.Range(c.Line, c.Col)
-					if s.Line == l && s.Col == col && end.Col == col+len(searchRunes) {
-						already = true
-						break
-					}
-				}
-				if already {
-					continue
-				}
-				sel := selection.Selection{Active: true, Anchor: selection.Position{Line: l, Col: col}}
-				e.Multi.AddWithSelection(l, col+len(searchRunes), sel)
-				e.syncFromMulti()
-				e.scrollViewport()
-				return
-			}
-		}
-	}
-}
-
-func (e *EditorPaneWidget) SelectAllOccurrences() {
-	if e.Selection == nil {
-		return
-	}
-	word := ""
-	if e.Selection.Active {
-		word = e.Selection.Text(e.Buf.Lines, e.Cursor.Line, e.Cursor.Col)
-	}
-	if word == "" {
-		e.selectWord(e.Cursor.Line, e.Cursor.Col)
-		if e.Selection.Active {
-			word = e.Selection.Text(e.Buf.Lines, e.Cursor.Line, e.Cursor.Col)
-		}
-	}
-	if word == "" {
-		return
-	}
-	e.multiSearchWord = word
-	e.ensureMulti()
-	e.syncToMulti()
-
-	searchRunes := []rune(word)
-	for line := 0; line < len(e.Buf.Lines); line++ {
-		runes := []rune(e.Buf.Lines[line])
-		for col := 0; col <= len(runes)-len(searchRunes); col++ {
-			if string(runes[col:col+len(searchRunes)]) == word {
-				sel := selection.Selection{Active: true, Anchor: selection.Position{Line: line, Col: col}}
-				e.Multi.AddWithSelection(line, col+len(searchRunes), sel)
-			}
-		}
-	}
-	e.syncFromMulti()
-}
-
-func (e *EditorPaneWidget) UndoLastCursor() {
-	if e.Multi == nil || !e.Multi.IsMulti() {
-		return
-	}
-	e.Multi.RemoveLast()
-	if !e.Multi.IsMulti() {
-		e.syncFromMulti()
-		e.Multi = nil
-		e.multiSearchWord = ""
-	} else {
-		e.syncFromMulti()
-	}
-	e.scrollViewport()
-}
-
-func (e *EditorPaneWidget) SplitSelectionToLines() {
-	if e.Selection == nil || !e.Selection.Active {
-		return
-	}
-	start, end := e.Selection.Range(e.Cursor.Line, e.Cursor.Col)
-	if start.Line == end.Line {
-		return
-	}
-	if start.Line >= len(e.Buf.Lines) || end.Line >= len(e.Buf.Lines) {
-		e.Selection.Clear()
-		return
-	}
-	// If the selection ends at column 0 of the last line,
-	// exclude that line (cursor sits at the start, nothing selected there)
-	if end.Col == 0 && end.Line > start.Line {
-		end.Line--
-		end.Col = len([]rune(e.Buf.Lines[end.Line]))
-	}
-	if start.Line == end.Line {
-		e.Selection.Clear()
-		return
-	}
-	e.ensureMulti()
-	e.syncToMulti()
-	// Place the primary cursor at the end of the first selected line
-	firstLineLen := len([]rune(e.Buf.Lines[start.Line]))
-	col := firstLineLen
-	e.Multi.Cursors[e.Multi.Primary] = multicursor.CursorState{
-		Line: start.Line,
-		Col:  col,
-	}
-	// Add a cursor at the end of each subsequent line in the selection
-	for line := start.Line + 1; line <= end.Line; line++ {
-		lineLen := len([]rune(e.Buf.Lines[line]))
-		c := lineLen
-		if line == end.Line && end.Col < c {
-			c = end.Col
-		}
-		e.Multi.Add(line, c)
-	}
-	e.Selection.Clear()
-	// Clear selections on all cursors
-	for i := range e.Multi.Cursors {
-		e.Multi.Cursors[i].Sel.Clear()
-	}
-	e.syncFromMulti()
-	e.scrollViewport()
-}
-
-// transformSelection replaces the selected text with the result of applying fn.
-// It preserves the selection after transformation.
-func (e *EditorPaneWidget) transformSelection(fn func(string) string) {
-	if e.Selection == nil || !e.Selection.Active {
-		return
-	}
-	text := e.Selection.Text(e.Buf.Lines, e.Cursor.Line, e.Cursor.Col)
-	if text == "" {
-		return
-	}
-	transformed := fn(text)
-	if transformed == text {
-		return
-	}
-
-	start, end := e.Selection.Range(e.Cursor.Line, e.Cursor.Col)
-
-	if e.Undo != nil {
-		e.Undo.BreakGroup()
-	}
-
-	oldLines := make([]string, end.Line-start.Line+1)
-	copy(oldLines, e.Buf.Lines[start.Line:end.Line+1])
-
-	tLines := strings.Split(transformed, "\n")
-	firstOld := []rune(oldLines[0])
-	startCol := min(start.Col, len(firstOld))
-	prefix := string(firstOld[:startCol])
-	lastOld := []rune(oldLines[len(oldLines)-1])
-	suffix := ""
-	if end.Col < len(lastOld) {
-		suffix = string(lastOld[end.Col:])
-	}
-
-	newLines := make([]string, len(tLines))
-	for i, tl := range tLines {
-		switch {
-		case len(tLines) == 1:
-			newLines[i] = prefix + tl + suffix
-		case i == 0:
-			newLines[i] = prefix + tl
-		case i == len(tLines)-1:
-			newLines[i] = tl + suffix
-		default:
-			newLines[i] = tl
-		}
-	}
-
-	cmd := &undo.ReplaceLinesCommand{
-		Start:    start.Line,
-		OldLines: oldLines,
-		NewLines: newLines,
-	}
-	e.exec(cmd)
-
-	newEndLine := start.Line + len(tLines) - 1
-	var newEndCol int
-	if len(tLines) == 1 {
-		newEndCol = startCol + len([]rune(tLines[0]))
-	} else {
-		newEndCol = len([]rune(tLines[len(tLines)-1]))
-	}
-
-	e.Selection.Start(start.Line, startCol)
-	e.Cursor.Line = newEndLine
-	e.Cursor.Col = newEndCol
-	e.clampCursor()
-	e.scrollViewport()
-}
-
-func (e *EditorPaneWidget) UpperCase() {
-	e.transformSelection(strings.ToUpper)
-}
-
-func (e *EditorPaneWidget) LowerCase() {
-	e.transformSelection(strings.ToLower)
-}
-
-func (e *EditorPaneWidget) TitleCase() {
-	e.transformSelection(func(s string) string {
-		runes := []rune(s)
-		inWord := false
-		for i, r := range runes {
-			if unicode.IsLetter(r) || unicode.IsDigit(r) {
-				if !inWord {
-					runes[i] = unicode.ToUpper(r)
-					inWord = true
-				} else {
-					runes[i] = unicode.ToLower(r)
-				}
-			} else if !(inWord && r == '\'') {
-				inWord = false
-			}
-		}
-		return string(runes)
-	})
 }
